backend/requirement.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-dotenv==1.0.1
selenium==4.25.0
webdriver-manager==4.0.2
